refactor: migrate Severity and violation-filter imports away from evaluators
- Severity: all consumers now import from schemas/rule-schemas
- computeFilterDecision/FilterDecision: all consumers now import from lint/violation-filter
- EvaluationType → ReviewType in 3 test files (import + usages)
