+++ src/cli/result-routing/check-routing.ts
@@ -1,6 +1,6 @@
 import type { RuleFile } from '../../rules/rule-loader';
 import type { RawCheckResult } from '../../prompts/schema';
-import { Severity } from '../../evaluators/types';
+import { Severity } from '../../schemas/rule-schemas';
 import { calculateCheckScore } from '../../scoring';
 import {
   buildRuleName,
